Misc cleanup, and move the cachedConsoleOutput function to ec2Instance.go
